before pagination changing

+++ src/main/java/dao/implementation/AdminDaoImpl.java
@@ -1,7 +1,6 @@
 package dao.implementation;
 
 import dao.AdminDAO;
-import dao.PersonalInfoDAO;
 import entity.Admin;
 import util.ConnectionPool;
 
@@ -11,19 +10,20 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Logger;
 
 
 public class AdminDaoImpl implements AdminDAO {
-
+    static final Logger logger = Logger.getLogger(String.valueOf(AdminDaoImpl.class));
     @Override
     public Admin get(long id) {
         Admin admin = new Admin();
         try(Connection con = ConnectionPool.getInstance().getConnection()) {
+            logger.info("try to find admin with id - " + id);
             String sql = "SELECT * FROM admin WHERE id = ?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, id);
             ResultSet rs = ps.executeQuery();
-            PersonalInfoDAO personalInfoDAO = new PersonalInfoDaoImpl();
             if(rs.next()){
                 admin.setId(rs.getInt("id"));
                 admin.setPersonId(rs.getInt("person_id"));
@@ -31,6 +31,7 @@ public class AdminDaoImpl implements AdminDAO {
             rs.close();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to find admin with id - " + id);
             throw new RuntimeException(e);
         }
         return admin;
@@ -42,6 +43,7 @@ public class AdminDaoImpl implements AdminDAO {
         String sql = "select * from admin";
         PreparedStatement ps;
         try(Connection con = ConnectionPool.getInstance().getConnection()) {
+            logger.info("try to get all admin in db");
             ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
             while (rs.next()){
@@ -52,6 +54,7 @@ public class AdminDaoImpl implements AdminDAO {
             }
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to get all admin");
             throw new RuntimeException(e);
         }
         return list;
@@ -59,7 +62,11 @@ public class AdminDaoImpl implements AdminDAO {
 
     @Override
     public void insert(Admin admin) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (admin == null) {
+            logger.warning("admin is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "INSERT INTO admin (id, person_id) VALUES(?, ?);";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, admin.getId());
@@ -67,13 +74,18 @@ public class AdminDaoImpl implements AdminDAO {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to insert admin with personalInfo id - " + admin.getPersonId());
             throw new RuntimeException(e);
         }
     }
 
     @Override
     public void update(Admin admin) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (admin == null) {
+            logger.warning("admin is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "UPDATE admin SET person_id = ? WHERE id = ?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, admin.getId());
@@ -81,19 +93,25 @@ public class AdminDaoImpl implements AdminDAO {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to update admin with id - " + admin.getId());
             throw new RuntimeException(e);
         }
     }
 
     @Override
     public void delete(Admin admin) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (admin == null) {
+            logger.warning("admin is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "DELETE FROM admin WHERE id = ?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, admin.getId());
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to delete admin with id - " + admin.getId());
             throw new RuntimeException(e);
         }
     }
@@ -111,6 +129,7 @@ public class AdminDaoImpl implements AdminDAO {
             rs.close();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to find admin by personalInfo id - " + id);
             throw new RuntimeException(e);
         }
         return false;
