+++ src/main/java/dao/implementation/AuthorDaoImpl.java
@@ -10,23 +10,29 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Logger;
 
 public class AuthorDaoImpl implements AuthorDAO {
+    static final Logger logger = Logger.getLogger(String.valueOf(AuthorDaoImpl.class));
 
     public long idFromName(String name, String nameUa) {
-        long id=0L;
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (name == null || nameUa == null) {
+            logger.warning("there is null in name");
+        }
+        long id = 0L;
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "SELECT * FROM author WHERE  author_name = ? and author_name_ua = ?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setString(1, name);
             ps.setString(2, nameUa);
             ResultSet rs = ps.executeQuery();
-            if(rs.next()){
+            if (rs.next()) {
                 id = rs.getInt("id");
             }
             rs.close();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to get id from name - " + name);
             throw new RuntimeException(e);
         }
         return id;
@@ -48,6 +54,7 @@ public class AuthorDaoImpl implements AuthorDAO {
             rs.close();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to get Author with id - " + id);
             throw new RuntimeException(e);
         }
         return author;
@@ -72,6 +79,8 @@ public class AuthorDaoImpl implements AuthorDAO {
             rs.close();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to get all Authors in db");
+
             throw new RuntimeException(e);
         }
         return list;
@@ -79,7 +88,11 @@ public class AuthorDaoImpl implements AuthorDAO {
 
     @Override
     public void insert(Author author) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (author == null) {
+            logger.warning("author is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "INSERT INTO author (id, author_name, author_name_ua) VALUES(?,?,?)";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, author.getId());
@@ -88,13 +101,18 @@ public class AuthorDaoImpl implements AuthorDAO {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to insert Author with name - " + author.getAuthorName());
             throw new RuntimeException(e);
         }
     }
 
     @Override
     public void update(Author author) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (author == null) {
+            logger.warning("author is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "UPDATE author SET author_name = ?, author_name_ua = ? WHERE id=?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setString(1, author.getAuthorName());
@@ -103,19 +121,25 @@ public class AuthorDaoImpl implements AuthorDAO {
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to update Author with id - " + author.getId());
             throw new RuntimeException(e);
         }
     }
 
     @Override
     public void delete(Author author) {
-        try(Connection con = ConnectionPool.getInstance().getConnection()) {
+        if (author == null) {
+            logger.warning("author is null");
+            return;
+        }
+        try (Connection con = ConnectionPool.getInstance().getConnection()) {
             String sql = "DELETE FROM author WHERE id = ?";
             PreparedStatement ps = con.prepareStatement(sql);
             ps.setLong(1, author.getId());
             ps.executeUpdate();
             ps.close();
         } catch (SQLException e) {
+            logger.warning("failed to delete Author with id - " + author.getId());
             throw new RuntimeException(e);
         }
     }
